ä161 Moved autoscroll into w2g file
