+++ src/javascript/configuration/configuration.ts
@@ -20,6 +20,7 @@ export const SETTINGS_playerAutoplayAfterFocusGain = 'playerAutoplayAfterFocusGa
 export const SETTINGS_w2gDisplayCharacterCounter = 'w2gDisplayCharacterCounter';
 export const SETTINGS_w2gAutotoggleHide = 'w2gAutotoggleHide';
 export const SETTINGS_w2gAutoscrollToUnseen = 'w2gAutoscrollToUnseen';
+
 class Configuration {
     settingsCache: Map<string, boolean>;
